requirements.txt
torch~=2.1.2+cu121
matplotlib~=3.7.2
seaborn~=0.13.2
torchvision~=0.16.2+cu121
scikit-learn~=1.3.2
numpy~=1.24.4
pillow~=10.4.0